+++ src/routes/changePage.js
@@ -1,0 +1,4 @@
+export const changePage = (page) => {
+  window.history.pushState({}, 'Loading...', page);
+  window.dispatchEvent(new PopStateEvent('popstate'));
+};
