+++ src/services/firebase.js
@@ -1,8 +1,6 @@
-/* eslint-disable import/named */
-/* eslint-disable import/no-cycle */
-/* eslint-disable no-alert */
-import { changePage } from '../router.js';
+import { changePage } from '../routes/changePage.js';
 
+/* eslint-disable no-alert */
 const getUser = () => firebase.auth().currentUser;
 
 const updateUser = async (name) => {
@@ -16,7 +14,7 @@ const signInWithGoogle = () => {
   const provider = new firebase.auth.GoogleAuthProvider();
   provider.addScope('profile');
   provider.addScope('email');
-  firebase.auth().signInWithPopup(provider)
+  return firebase.auth().signInWithPopup(provider)
     .then((result) => {
       changePage('/');
       console.log(result);
@@ -30,7 +28,7 @@ const signUpWithGoogle = () => {
   const providerSignUp = new firebase.auth.GoogleAuthProvider();
   providerSignUp.addScope('profile');
   providerSignUp.addScope('email');
-  firebase.auth().signInWithPopup(providerSignUp)
+  return firebase.auth().signInWithPopup(providerSignUp)
     .then((result) => {
       changePage('/');
       console.log(result);
